# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Centros",
        "description": "Documentação da API de Centros, incluindo operações de CRUD e relacionamentos.",
        "version": "1.1.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API principal"
        }
    ],
    "paths": {
...
        "/pre-inscricoes": {
            "get": {
                "tags": [
                    "Pré-inscrições"
                ],
                "summary": "Listar todas as pré-inscrições",
                "operationId": "0c5003dd459f7f6da31baef278d72314",
                "responses": {
                    "200": {
                        "description": "Lista de pré-inscrições",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PreInscricao"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pré-inscrições"
                ],
                "summary": "Realizar uma nova pré-inscrição",
                "operationId": "acdf02d2794da64d6e124ab9bc209279",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PreInscricaoInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pré-inscrição realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PreInscricao"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/pre-inscricoes/{id}": {
            "get": {
                "tags": [
                    "Pré-inscrições"
                ],
                "summary": "Buscar pré-inscrição por ID",
                "operationId": "a23c806c7228ecb74e0cab5ebbf48271",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pré-inscrição encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PreInscricao"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pré-inscrição não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pré-inscrições"
                ],
                "summary": "Atualizar status da pré-inscrição (apenas status)",
                "operationId": "6578938d922d5c5dc7fb27bafd41d26b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PreInscricaoStatusInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PreInscricao"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pré-inscrição não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pré-inscrições"
                ],
                "summary": "Deletar pré-inscrição",
                "operationId": "9d089f4f9b6046b11bc19d8251357341",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pré-inscrição deletada"
                    },
                    "404": {
                        "description": "Pré-inscrição não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PreInscricao": {
                "required": [
                    "id",
                    "curso_id",
                    "centro_id",
                    "nome_completo",
                    "contactos",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "curso_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "centro_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "horario_id": {
                        "type": "integer",
                        "example": 4,
                        "nullable": true
                    },
                    "nome_completo": {
                        "type": "string",
                        "example": "Maria dos Santos"
                    },
                    "contactos": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "maria@email.com",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pendente",
                            "confirmado",
                            "cancelado"
                        ],
                        "example": "pendente"
                    },
                    "observacoes": {
                        "type": "string",
                        "example": "Precisa de material especial",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-31T08:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-31T08:00:00Z"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "curso_id": 2,
                    "centro_id": 3,
                    "horario_id": 4,
                    "nome_completo": "Maria dos Santos",
                    "contactos": [
                        "923456789",
                        "912345678"
                    ],
                    "email": "maria@email.com",
                    "status": "pendente",
                    "observacoes": "Precisa de material especial",
                    "created_at": "2025-07-31T08:00:00Z",
                    "updated_at": "2025-07-31T08:00:00Z"
                }
            },
            "PreInscricaoInput": {
                "required": [
                    "curso_id",
                    "centro_id",
                    "nome_completo",
                    "contactos"
                ],
                "properties": {
                    "curso_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "centro_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "horario_id": {
                        "type": "integer",
                        "example": 4,
                        "nullable": true
                    },
                    "nome_completo": {
                        "type": "string",
                        "example": "Maria dos Santos"
                    },
                    "contactos": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "maria@email.com",
                        "nullable": true
                    },
                    "observacoes": {
                        "type": "string",
                        "example": "Precisa de material especial",
                        "nullable": true
                    }
                },
                "type": "object",
                "example": {
                    "curso_id": 2,
                    "centro_id": 3,
                    "horario_id": 4,
                    "nome_completo": "Maria dos Santos",
                    "contactos": [
                        "923456789",
                        "912345678"
                    ],
                    "email": "maria@email.com",
                    "observacoes": "Precisa de material especial"
                }
            },
            "PreInscricaoStatusInput": {
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "pendente",
                            "confirmado",
                            "cancelado"
                        ],
                        "example": "pendente"
                    }
                },
                "type": "object",
                "example": {
                    "status": "pendente"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}